Rearrange files, add specs for MapUtils

+++ src/main/scala/heist/Types.scala
@@ -2,4 +2,4 @@ package heist
 
 final case class BankCoordinate(x: Double, y: Double)
 
-final case class BankMap(width: Float, detectors: List[BankCoordinate])
+final case class BankMap(width: Double, detectors: List[BankCoordinate])
